requirements.txt
pytest==8.0.2
allure-pytest==2.13.2
selene>=2.0.b8
python-dotenv==1.0.1
